move useeffect into contract component
