+++ src/components/profile/contract/Contract.tsx
@@ -1,19 +1,62 @@
-import { Dispatch, SetStateAction, useState } from "react";
-import { ContractData } from "../../../types";
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
+import { BLANK_CONTRACT, ContractData } from "../../../types";
 import { IconButton, Modal, ModalClose, ModalDialog, Stack, Typography } from "@mui/joy";
 import { ArrowSquareOut } from "@phosphor-icons/react";
 import ContractEdit from "./ContractEdit";
 import ContractText from "./ContractText";
+import { supabase } from "../../../supabaseClient";
+import { Session } from "@supabase/supabase-js";
+import { useOutletContext } from "react-router-dom";
 
-interface ContractProps{
-  contract: ContractData;
-  setContract: Dispatch<SetStateAction<ContractData>>;
-  lastUpdated: Date | null;
-  onSave: () => Promise<void>;
-}
-
-export default function Contract({ contract, setContract, lastUpdated, onSave }: ContractProps) {
+export default function Contract() {
   const [open, setOpen] = useState(false);
+  const [contract, setContract] = useState<ContractData>(BLANK_CONTRACT);
+  const [lastUpdated, setLastUpdated] = useState<Date | null>(null);
+  const [loading, setLoading] = useState(false);
+
+  const { session } = useOutletContext<{ session: Session | null }>();
+
+  useEffect(() => {
+    (async function fetchContract() {
+      const {data, error} = await supabase
+      .from('contracts')
+      .select('data, updated_at')
+      .eq('profile_id', session?.user.id)
+      .order('updated_at', {ascending:false})
+      .limit(1)
+      .single()
+      
+      if (error) {
+        console.error(error);
+      } else {
+        setContract(data.data as ContractData);
+        setLastUpdated(new Date(data.updated_at));
+      }
+    })();
+  }, [session])
+
+  async function handleContractSave() {
+    const now = new Date()
+    const {error, data } = await supabase
+    .from('contracts')
+    .upsert({
+      profile_id: session?.user.id,
+      data: contract,
+      updated_at: now
+    })
+    .select('data, updated_at')
+    .single()
+    
+    if (error) {
+      console.error(error);
+    }
+    else {
+      setContract(data.data as ContractData);
+      setLastUpdated(new Date(data.updated_at));
+    }
+  
+    setLoading(false);
+  }
 
   return (
     <>
@@ -34,7 +77,7 @@ export default function Contract({ contract, setContract, lastUpdated, onSave }:
         </Typography>
       </Stack>
 
-      <ContractModal open={open} setOpen={setOpen} contract={contract} lastUpdated={lastUpdated} onSave={onSave} setContract={setContract}/>
+      <ContractModal open={open} setOpen={setOpen} contract={contract} lastUpdated={lastUpdated} onSave={handleContractSave} setContract={setContract}/>
     </>
   )
 }
